extract and explicitly mock timezone
